+++ src/utils/logout.ts
@@ -2,12 +2,12 @@ import { getAccountProvider, getProviderType } from 'providers';
 import { logoutAction } from 'reduxStore/commonActions';
 import { store } from 'reduxStore/store';
 import { LoginMethodsEnum } from 'types';
-import { getIsLoggedIn } from 'utils/getIsLoggedIn';
+import { matchPath } from '../wrappers/AuthenticatedRoutesWrapper/helpers/matchPath';
 import { getAddress, getWebviewToken } from './account';
 import { preventRedirects, safeRedirect } from './redirect';
 import { storage } from './storage';
 import { localStorageKeys } from './storage/local';
-import { addOriginToLocationPath } from './window';
+import { addOriginToLocationPath, getWindowLocation } from './window';
 
 const broadcastLogoutAcrossTabs = (address: string) => {
   const storedData = storage.local.getItem(localStorageKeys.logoutEvent);
@@ -22,29 +22,21 @@ const broadcastLogoutAcrossTabs = (address: string) => {
     data: address,
     expires: 0
   });
+
   storage.local.removeItem(localStorageKeys.logoutEvent);
 };
 
 export async function logout(
   callbackUrl?: string,
   onRedirect?: (callbackUrl?: string) => void,
-  shouldAttemptRelogin = Boolean(getWebviewToken())
+  shouldAttemptReLogin = Boolean(getWebviewToken())
 ) {
-  const isLoggedIn = getIsLoggedIn();
   const provider = getAccountProvider();
   const providerType = getProviderType(provider);
   const isWalletProvider = providerType === LoginMethodsEnum.wallet;
 
-  if (shouldAttemptRelogin && provider?.relogin != null) {
+  if (shouldAttemptReLogin && provider?.relogin != null) {
     return await provider.relogin();
-  }
-
-  if (isWalletProvider) {
-    if (!isLoggedIn) {
-      return;
-    }
-
-    preventRedirects();
   }
 
   const url = addOriginToLocationPath(callbackUrl);
@@ -55,10 +47,18 @@ export async function logout(
     broadcastLogoutAcrossTabs(address);
   } catch (err) {
     console.error('error fetching logout address', err);
-    return redirectToCallbackUrl(url, onRedirect);
   }
 
-  if (!provider || providerType === LoginMethodsEnum.none) {
+  const location = getWindowLocation();
+  const callbackPathname = new URL(url).pathname;
+
+  // Prevent page redirect if
+  // the logout callbackURL is equal to the current URL
+  if (matchPath(location.pathname, callbackPathname)) {
+    preventRedirects();
+  }
+
+  if (!provider?.isInitialized?.()) {
     return redirectToCallbackUrl(url, onRedirect);
   }
 
@@ -76,10 +76,9 @@ export async function logout(
 
 function redirectToCallbackUrl(
   callbackUrl?: string,
-  onRedirect?: (callbackUrl?: string) => void,
-  isWalletProvider?: boolean
+  onRedirect?: (callbackUrl?: string) => void
 ) {
-  if (callbackUrl && !isWalletProvider) {
+  if (callbackUrl) {
     if (typeof onRedirect === 'function') {
       onRedirect(callbackUrl);
     } else {